requirements.txt
absl-py==1.4.0
adabelief-pytorch==0.2.1
cachetools==4.2.4
charset-normalizer==2.0.12
cloudpickle==2.2.1
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
dgl-cu110==0.6.1
dgllife==0.2.9
future==1.0.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
graphviz==0.19.1
grpcio==1.48.2
hyperopt==0.2.7
idna==3.10
imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.3.1
littleutils==0.2.2
llvmlite==0.36.0
Markdown==3.3.7
matplotlib==3.3.4
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11
oauthlib==3.2.2
ogb==1.3.6
outdated==0.2.2
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
psutil==7.1.0
py4j==0.10.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
rdkit-pypi==2021.9.4
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.17.0
tabulate==0.8.10
threadpoolctl==3.1.0
torch==1.10.1
torchsummary==1.5.1
torchvision==0.11.2
torchviz==0.0.3
tqdm==4.64.1
typing-extensions==4.1.1
umap-learn==0.5.7
urllib3==1.26.20
Werkzeug==2.0.3
zipp==3.6.0
